refactor buildSettings
