+++ Sources/XcodeProjectManagement/ProjectSettings/XcodeBuildSettingsForIndex.swift
@@ -1,0 +1,89 @@
+public typealias XcodeBuildSettingsForIndex = [String: [String: XcodeFileBuildSettingInfo]]
+
+/// Represents a build setting value that can be a string or array of strings
+public enum BuildSettingValue: Codable {
+    case string(String)
+    case array([String])
+
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.singleValueContainer()
+
+        if let stringValue = try? container.decode(String.self) {
+            self = .string(stringValue)
+        } else if let arrayValue = try? container.decode([String].self) {
+            self = .array(arrayValue)
+        } else {
+            throw DecodingError.typeMismatch(
+                BuildSettingValue.self,
+                DecodingError.Context(codingPath: decoder.codingPath, debugDescription: "Expected String or [String]")
+            )
+        }
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.singleValueContainer()
+        switch self {
+        case let .string(value):
+            try container.encode(value)
+        case let .array(values):
+            try container.encode(values)
+        }
+    }
+
+    /// Convert to string representation
+    public var stringValue: String {
+        switch self {
+        case let .string(value):
+            value
+        case let .array(values):
+            values.joined(separator: " ")
+        }
+    }
+}
+
+public struct XcodeFileBuildSettingInfo: Codable, Sendable {
+    public var assetSymbolIndexPath: String?
+    public var languageDialect: XcodeLanguageDialect?
+    public var outputFilePath: String?
+    public var swiftASTBuiltProductsDir: String?
+    public var swiftASTCommandArguments: [String]?
+    public var swiftASTModuleName: String?
+    public var toolchains: [String]?
+
+    private enum CodingKeys: String, CodingKey {
+        case assetSymbolIndexPath
+        case languageDialect = "LanguageDialect" // Note: Xcode uses capital 'L'
+        case outputFilePath
+        case swiftASTBuiltProductsDir
+        case swiftASTCommandArguments
+        case swiftASTModuleName
+        case toolchains
+    }
+
+    public init(
+        assetSymbolIndexPath: String? = nil,
+        languageDialect: XcodeLanguageDialect? = nil,
+        outputFilePath: String? = nil,
+        swiftASTBuiltProductsDir: String? = nil,
+        swiftASTCommandArguments: [String]? = nil,
+        swiftASTModuleName: String? = nil,
+        toolchains: [String]? = nil
+    ) {
+        self.assetSymbolIndexPath = assetSymbolIndexPath
+        self.languageDialect = languageDialect
+        self.outputFilePath = outputFilePath
+        self.swiftASTBuiltProductsDir = swiftASTBuiltProductsDir
+        self.swiftASTCommandArguments = swiftASTCommandArguments
+        self.swiftASTModuleName = swiftASTModuleName
+        self.toolchains = toolchains
+    }
+}
+
+public extension XcodeBuildSettingsForIndex {
+    func fileBuildInfo(for target: String, fileName: String) -> XcodeFileBuildSettingInfo? {
+        guard let targetBuildSettings = self[target] else {
+            return nil
+        }
+        return targetBuildSettings[fileName]
+    }
+}
