+++ Sources/XcodeProjectManagement/ProjectSettings/XcodeBuildSettingsForIndexFixture.swift
