+++ Sources/XcodeProjectManagement/ProjectSettings/XcodeSettingsLoader.swift
@@ -1,0 +1,61 @@
+//
+//  XcodeSettingsLoader.swift
+//
+//  Copyright © 2024 Wang Lun.
+//
+
+import Foundation
+import Logger
+
+public actor XcodeSettingsLoader {
+    let commandBuilder: XcodeBuildCommandBuilder
+    private let toolchain: XcodeToolchain
+    let jsonDecoder = JSONDecoder()
+
+    public init(commandBuilder: XcodeBuildCommandBuilder, toolchain: XcodeToolchain) {
+        self.commandBuilder = commandBuilder
+        self.toolchain = toolchain
+    }
+
+    public func loadIndexingPaths(
+        buildSettingsList: [XcodeBuildSettings]
+    ) async throws -> (indexStoreURL: URL, indexDatabaseURL: URL) {
+        guard let settings = buildSettingsList.first?.buildSettings else {
+            throw XcodeProjectError.invalidConfig("No build settings found")
+        }
+
+        guard let buildFolderPath = settings["BUILD_DIR"] else {
+            throw XcodeProjectError.invalidConfig("BUILD_DIR not found in build settings")
+        }
+
+        let outputFolder = URL(fileURLWithPath: buildFolderPath)
+            .deletingLastPathComponent()
+            .deletingLastPathComponent()
+
+        let indexStoreURL = outputFolder.appendingPathComponent("Index.noIndex/DataStore")
+        let indexDatabaseURL = outputFolder.appendingPathComponent("IndexDatabase.noIndex")
+
+        do {
+            if !FileManager.default.fileExists(atPath: indexDatabaseURL.path) {
+                try FileManager.default.createDirectory(at: indexDatabaseURL, withIntermediateDirectories: true)
+            }
+        } catch {
+            throw XcodeProjectError.invalidConfig("Failed to create index database directory: \(error)")
+        }
+
+        return (indexStoreURL: indexStoreURL, indexDatabaseURL: indexDatabaseURL)
+    }
+
+    public func runXcodeBuild(
+        arguments: [String],
+        workingDirectory: URL
+    ) async throws -> String? {
+        let result = try await toolchain.executeXcodeBuild(
+            arguments: arguments,
+            workingDirectory: workingDirectory
+        )
+
+        // 返回空字符串时返回nil
+        return result.output.isEmpty ? nil : result.output
+    }
+}
